+++ src/tests/Workflows/hallo-09.scriptbook.ps1
@@ -1,0 +1,16 @@
+[CmdletBinding(SupportsShouldProcess = $True)]
+Param()
+
+Set-Location $PSScriptRoot
+
+Import-Module ../../Scriptbook/Scriptbook.psm1 -Force -Args @{ Quiet = $true }
+
+Action Hello {
+    Write-Info "Hello"
+}
+
+Action GoodBy {
+    Write-Info "GoodBy"
+}
+
+Start-Workflow -Name HelloWorkflow -WhatIf:$WhatIfPreference -NoLogging -NoReport