+++ docs/release-notes.md
@@ -4,6 +4,16 @@
 ![PowerShell Gallery](https://img.shields.io/powershellgallery/dt/Scriptbook.svg?label=PSGallery%20Downloads&logo=PowerShell&style=flat-square)
 [![Build Status](https://dev.azure.com/tedon/TD.Deploy/_apis/build/status/ehagen.Scriptbook?branchName=master)](https://dev.azure.com/tedon/TD.Deploy/_build/latest?definitionId=52&branchName=master)
 
+## [0.6.4] 2022-10-22
+
+```plain
+### Fixed
+- Added parameters merge support by default, merges parameter values defined in script with values loaded from file
+- Fixed NoLogging parameter of Start-Workflow in combination with Verbose usage
+- Added secrets scanning during build of module
+
+```
+
 ## [0.6.3]
 
 ```plain